+++ .vscode/settings.json
@@ -1,3 +1,9 @@
 {
-    "typescript.format.semicolons": "insert"
+    "typescript.format.semicolons": "insert",
+    "typescript.inlayHints.enumMemberValues.enabled": true,
+    "typescript.implementationsCodeLens.enabled": true,
+    "javascript.referencesCodeLens.enabled": true,
+    "javascript.referencesCodeLens.showOnAllFunctions": true,
+    "typescript.referencesCodeLens.enabled": true,
+    "typescript.referencesCodeLens.showOnAllFunctions": true
 }
